Port FTE's WebRTC code over to QSS.
`sv_port_rtc /foo` on the server and `connect /foo` on the client can be used to host games a little more easily, using (short?) names instead of ip addresses.
This also gives support for both dtls://ip:port?fp=FINGERPRINT and wss:// schemes (on linux, with gnutls). Use -pubkey PATHTOFULLCHAIN.pem and -privkey PATHTOPRIVKEY.pem (as eg generated by letsencrypt) to use certificates that other people might be able to trust. It is possible to generate self-signed certificates with -certhost YOURDNS, but browsers will still reject it. Note that webrtc uses temporary certificates and thus does not use these args.
This stuff is compatible with FTE, as well as able to act as a server for netquake.io - it is recommended to just point your browser at the game's wss socket as if it were an https server - it'll provide redirect links and a way for you to more easily add any requisite certificate exceptions/pinning to your browser.

The new code is disabled by default for now, but can be enabled with -useice. The new code listens on $port+1000, althouth the old net_dgram.c code should probably be disabled instead of using redundant sockets.